requirements.txt
dm_control==1.0.10
fbm==0.3.0
matplotlib==3.5.2
ml_collections==0.1.1
numpy==1.21.5
pandas==1.3.5
Pillow==9.4.0
PyYAML==6.0
scikit_learn==1.2.1
scipy==1.7.3
seaborn==0.10.0
tqdm==4.64.0
yfinance==0.2.12
chardet
gdown